Moved gfx low level interface into sdl2_sys crate
Progresses on #647
