+++ sdl2-sys/Cargo.toml
@@ -53,3 +53,4 @@ bundled = ["cmake", "reqwest", "tar", "flate2"]
 mixer = []
 image = []
 ttf = []
+gfx = []
