+++ sdl2-sys/src/gfx/framerate.rs
@@ -1,0 +1,18 @@
+/* automatically generated by rust-bindgen */
+
+use ::std::os::raw::*;
+#[repr(C)]
+pub struct FPSmanager {
+    pub framecount: c_uint,
+    pub rateticks: c_float,
+    pub baseticks: c_uint,
+    pub lastticks: c_uint,
+    pub rate: c_uint,
+}
+extern "C" {
+    pub fn SDL_initFramerate(manager: *mut FPSmanager);
+    pub fn SDL_setFramerate(manager: *mut FPSmanager, rate: c_uint) -> c_int;
+    pub fn SDL_getFramerate(manager: *mut FPSmanager) -> c_int;
+    pub fn SDL_getFramecount(manager: *mut FPSmanager) -> c_int;
+    pub fn SDL_framerateDelay(manager: *mut FPSmanager) -> c_uint;
+}
