+++ sdl2-sys/src/gfx/mod.rs
@@ -1,0 +1,4 @@
+pub mod primitives;
+pub mod framerate;
+pub mod imagefilter;
+pub mod rotozoom;
