+++ sdl2-sys/src/gfx/primitives.rs
@@ -1,0 +1,494 @@
+/* automatically generated by rust-bindgen */
+
+use ::SDL_Renderer;
+use ::SDL_Surface;
+use ::std::os::raw::*;
+
+extern "C" {
+    pub fn pixelColor(renderer: *const SDL_Renderer,
+                      x: c_short,
+                      y: c_short,
+                      color: c_uint)
+                      -> c_int;
+    pub fn pixelRGBA(renderer: *const SDL_Renderer,
+                     x: c_short,
+                     y: c_short,
+                     r: c_uchar,
+                     g: c_uchar,
+                     b: c_uchar,
+                     a: c_uchar)
+                     -> c_int;
+    pub fn hlineColor(renderer: *const SDL_Renderer,
+                      x1: c_short,
+                      x2: c_short,
+                      y: c_short,
+                      color: c_uint)
+                      -> c_int;
+    pub fn hlineRGBA(renderer: *const SDL_Renderer,
+                     x1: c_short,
+                     x2: c_short,
+                     y: c_short,
+                     r: c_uchar,
+                     g: c_uchar,
+                     b: c_uchar,
+                     a: c_uchar)
+                     -> c_int;
+    pub fn vlineColor(renderer: *const SDL_Renderer,
+                      x: c_short,
+                      y1: c_short,
+                      y2: c_short,
+                      color: c_uint)
+                      -> c_int;
+    pub fn vlineRGBA(renderer: *const SDL_Renderer,
+                     x: c_short,
+                     y1: c_short,
+                     y2: c_short,
+                     r: c_uchar,
+                     g: c_uchar,
+                     b: c_uchar,
+                     a: c_uchar)
+                     -> c_int;
+    pub fn rectangleColor(renderer: *const SDL_Renderer,
+                          x1: c_short,
+                          y1: c_short,
+                          x2: c_short,
+                          y2: c_short,
+                          color: c_uint)
+                          -> c_int;
+    pub fn rectangleRGBA(renderer: *const SDL_Renderer,
+                         x1: c_short,
+                         y1: c_short,
+                         x2: c_short,
+                         y2: c_short,
+                         r: c_uchar,
+                         g: c_uchar,
+                         b: c_uchar,
+                         a: c_uchar)
+                         -> c_int;
+    pub fn roundedRectangleColor(renderer: *const SDL_Renderer,
+                                 x1: c_short,
+                                 y1: c_short,
+                                 x2: c_short,
+                                 y2: c_short,
+                                 rad: c_short,
+                                 color: c_uint)
+                                 -> c_int;
+    pub fn roundedRectangleRGBA(renderer: *const SDL_Renderer,
+                                x1: c_short,
+                                y1: c_short,
+                                x2: c_short,
+                                y2: c_short,
+                                rad: c_short,
+                                r: c_uchar,
+                                g: c_uchar,
+                                b: c_uchar,
+                                a: c_uchar)
+                                -> c_int;
+    pub fn boxColor(renderer: *const SDL_Renderer,
+                    x1: c_short,
+                    y1: c_short,
+                    x2: c_short,
+                    y2: c_short,
+                    color: c_uint)
+                    -> c_int;
+    pub fn boxRGBA(renderer: *const SDL_Renderer,
+                   x1: c_short,
+                   y1: c_short,
+                   x2: c_short,
+                   y2: c_short,
+                   r: c_uchar,
+                   g: c_uchar,
+                   b: c_uchar,
+                   a: c_uchar)
+                   -> c_int;
+    pub fn roundedBoxColor(renderer: *const SDL_Renderer,
+                           x1: c_short,
+                           y1: c_short,
+                           x2: c_short,
+                           y2: c_short,
+                           rad: c_short,
+                           color: c_uint)
+                           -> c_int;
+    pub fn roundedBoxRGBA(renderer: *const SDL_Renderer,
+                          x1: c_short,
+                          y1: c_short,
+                          x2: c_short,
+                          y2: c_short,
+                          rad: c_short,
+                          r: c_uchar,
+                          g: c_uchar,
+                          b: c_uchar,
+                          a: c_uchar)
+                          -> c_int;
+    pub fn lineColor(renderer: *const SDL_Renderer,
+                     x1: c_short,
+                     y1: c_short,
+                     x2: c_short,
+                     y2: c_short,
+                     color: c_uint)
+                     -> c_int;
+    pub fn lineRGBA(renderer: *const SDL_Renderer,
+                    x1: c_short,
+                    y1: c_short,
+                    x2: c_short,
+                    y2: c_short,
+                    r: c_uchar,
+                    g: c_uchar,
+                    b: c_uchar,
+                    a: c_uchar)
+                    -> c_int;
+    pub fn aalineColor(renderer: *const SDL_Renderer,
+                       x1: c_short,
+                       y1: c_short,
+                       x2: c_short,
+                       y2: c_short,
+                       color: c_uint)
+                       -> c_int;
+    pub fn aalineRGBA(renderer: *const SDL_Renderer,
+                      x1: c_short,
+                      y1: c_short,
+                      x2: c_short,
+                      y2: c_short,
+                      r: c_uchar,
+                      g: c_uchar,
+                      b: c_uchar,
+                      a: c_uchar)
+                      -> c_int;
+    pub fn thickLineColor(renderer: *const SDL_Renderer,
+                          x1: c_short,
+                          y1: c_short,
+                          x2: c_short,
+                          y2: c_short,
+                          width: c_uchar,
+                          color: c_uint)
+                          -> c_int;
+    pub fn thickLineRGBA(renderer: *const SDL_Renderer,
+                         x1: c_short,
+                         y1: c_short,
+                         x2: c_short,
+                         y2: c_short,
+                         width: c_uchar,
+                         r: c_uchar,
+                         g: c_uchar,
+                         b: c_uchar,
+                         a: c_uchar)
+                         -> c_int;
+    pub fn circleColor(renderer: *const SDL_Renderer,
+                       x: c_short,
+                       y: c_short,
+                       rad: c_short,
+                       color: c_uint)
+                       -> c_int;
+    pub fn circleRGBA(renderer: *const SDL_Renderer,
+                      x: c_short,
+                      y: c_short,
+                      rad: c_short,
+                      r: c_uchar,
+                      g: c_uchar,
+                      b: c_uchar,
+                      a: c_uchar)
+                      -> c_int;
+    pub fn arcColor(renderer: *const SDL_Renderer,
+                    x: c_short,
+                    y: c_short,
+                    rad: c_short,
+                    start: c_short,
+                    end: c_short,
+                    color: c_uint)
+                    -> c_int;
+    pub fn arcRGBA(renderer: *const SDL_Renderer,
+                   x: c_short,
+                   y: c_short,
+                   rad: c_short,
+                   start: c_short,
+                   end: c_short,
+                   r: c_uchar,
+                   g: c_uchar,
+                   b: c_uchar,
+                   a: c_uchar)
+                   -> c_int;
+    pub fn aacircleColor(renderer: *const SDL_Renderer,
+                         x: c_short,
+                         y: c_short,
+                         rad: c_short,
+                         color: c_uint)
+                         -> c_int;
+    pub fn aacircleRGBA(renderer: *const SDL_Renderer,
+                        x: c_short,
+                        y: c_short,
+                        rad: c_short,
+                        r: c_uchar,
+                        g: c_uchar,
+                        b: c_uchar,
+                        a: c_uchar)
+                        -> c_int;
+    pub fn filledCircleColor(renderer: *const SDL_Renderer,
+                             x: c_short,
+                             y: c_short,
+                             r: c_short,
+                             color: c_uint)
+                             -> c_int;
+    pub fn filledCircleRGBA(renderer: *const SDL_Renderer,
+                            x: c_short,
+                            y: c_short,
+                            rad: c_short,
+                            r: c_uchar,
+                            g: c_uchar,
+                            b: c_uchar,
+                            a: c_uchar)
+                            -> c_int;
+    pub fn ellipseColor(renderer: *const SDL_Renderer,
+                        x: c_short,
+                        y: c_short,
+                        rx: c_short,
+                        ry: c_short,
+                        color: c_uint)
+                        -> c_int;
+    pub fn ellipseRGBA(renderer: *const SDL_Renderer,
+                       x: c_short,
+                       y: c_short,
+                       rx: c_short,
+                       ry: c_short,
+                       r: c_uchar,
+                       g: c_uchar,
+                       b: c_uchar,
+                       a: c_uchar)
+                       -> c_int;
+    pub fn aaellipseColor(renderer: *const SDL_Renderer,
+                          x: c_short,
+                          y: c_short,
+                          rx: c_short,
+                          ry: c_short,
+                          color: c_uint)
+                          -> c_int;
+    pub fn aaellipseRGBA(renderer: *const SDL_Renderer,
+                         x: c_short,
+                         y: c_short,
+                         rx: c_short,
+                         ry: c_short,
+                         r: c_uchar,
+                         g: c_uchar,
+                         b: c_uchar,
+                         a: c_uchar)
+                         -> c_int;
+    pub fn filledEllipseColor(renderer: *const SDL_Renderer,
+                              x: c_short,
+                              y: c_short,
+                              rx: c_short,
+                              ry: c_short,
+                              color: c_uint)
+                              -> c_int;
+    pub fn filledEllipseRGBA(renderer: *const SDL_Renderer,
+                             x: c_short,
+                             y: c_short,
+                             rx: c_short,
+                             ry: c_short,
+                             r: c_uchar,
+                             g: c_uchar,
+                             b: c_uchar,
+                             a: c_uchar)
+                             -> c_int;
+    pub fn pieColor(renderer: *const SDL_Renderer,
+                    x: c_short,
+                    y: c_short,
+                    rad: c_short,
+                    start: c_short,
+                    end: c_short,
+                    color: c_uint)
+                    -> c_int;
+    pub fn pieRGBA(renderer: *const SDL_Renderer,
+                   x: c_short,
+                   y: c_short,
+                   rad: c_short,
+                   start: c_short,
+                   end: c_short,
+                   r: c_uchar,
+                   g: c_uchar,
+                   b: c_uchar,
+                   a: c_uchar)
+                   -> c_int;
+    pub fn filledPieColor(renderer: *const SDL_Renderer,
+                          x: c_short,
+                          y: c_short,
+                          rad: c_short,
+                          start: c_short,
+                          end: c_short,
+                          color: c_uint)
+                          -> c_int;
+    pub fn filledPieRGBA(renderer: *const SDL_Renderer,
+                         x: c_short,
+                         y: c_short,
+                         rad: c_short,
+                         start: c_short,
+                         end: c_short,
+                         r: c_uchar,
+                         g: c_uchar,
+                         b: c_uchar,
+                         a: c_uchar)
+                         -> c_int;
+    pub fn trigonColor(renderer: *const SDL_Renderer,
+                       x1: c_short,
+                       y1: c_short,
+                       x2: c_short,
+                       y2: c_short,
+                       x3: c_short,
+                       y3: c_short,
+                       color: c_uint)
+                       -> c_int;
+    pub fn trigonRGBA(renderer: *const SDL_Renderer,
+                      x1: c_short,
+                      y1: c_short,
+                      x2: c_short,
+                      y2: c_short,
+                      x3: c_short,
+                      y3: c_short,
+                      r: c_uchar,
+                      g: c_uchar,
+                      b: c_uchar,
+                      a: c_uchar)
+                      -> c_int;
+    pub fn aatrigonColor(renderer: *const SDL_Renderer,
+                         x1: c_short,
+                         y1: c_short,
+                         x2: c_short,
+                         y2: c_short,
+                         x3: c_short,
+                         y3: c_short,
+                         color: c_uint)
+                         -> c_int;
+    pub fn aatrigonRGBA(renderer: *const SDL_Renderer,
+                        x1: c_short,
+                        y1: c_short,
+                        x2: c_short,
+                        y2: c_short,
+                        x3: c_short,
+                        y3: c_short,
+                        r: c_uchar,
+                        g: c_uchar,
+                        b: c_uchar,
+                        a: c_uchar)
+                        -> c_int;
+    pub fn filledTrigonColor(renderer: *const SDL_Renderer,
+                             x1: c_short,
+                             y1: c_short,
+                             x2: c_short,
+                             y2: c_short,
+                             x3: c_short,
+                             y3: c_short,
+                             color: c_uint)
+                             -> c_int;
+    pub fn filledTrigonRGBA(renderer: *const SDL_Renderer,
+                            x1: c_short,
+                            y1: c_short,
+                            x2: c_short,
+                            y2: c_short,
+                            x3: c_short,
+                            y3: c_short,
+                            r: c_uchar,
+                            g: c_uchar,
+                            b: c_uchar,
+                            a: c_uchar)
+                            -> c_int;
+    pub fn polygonColor(renderer: *const SDL_Renderer,
+                        vx: *const c_short,
+                        vy: *const c_short,
+                        n: c_int,
+                        color: c_uint)
+                        -> c_int;
+    pub fn polygonRGBA(renderer: *const SDL_Renderer,
+                       vx: *const c_short,
+                       vy: *const c_short,
+                       n: c_int,
+                       r: c_uchar,
+                       g: c_uchar,
+                       b: c_uchar,
+                       a: c_uchar)
+                       -> c_int;
+    pub fn aapolygonColor(renderer: *const SDL_Renderer,
+                          vx: *const c_short,
+                          vy: *const c_short,
+                          n: c_int,
+                          color: c_uint)
+                          -> c_int;
+    pub fn aapolygonRGBA(renderer: *const SDL_Renderer,
+                         vx: *const c_short,
+                         vy: *const c_short,
+                         n: c_int,
+                         r: c_uchar,
+                         g: c_uchar,
+                         b: c_uchar,
+                         a: c_uchar)
+                         -> c_int;
+    pub fn filledPolygonColor(renderer: *const SDL_Renderer,
+                              vx: *const c_short,
+                              vy: *const c_short,
+                              n: c_int,
+                              color: c_uint)
+                              -> c_int;
+    pub fn filledPolygonRGBA(renderer: *const SDL_Renderer,
+                             vx: *const c_short,
+                             vy: *const c_short,
+                             n: c_int,
+                             r: c_uchar,
+                             g: c_uchar,
+                             b: c_uchar,
+                             a: c_uchar)
+                             -> c_int;
+    pub fn texturedPolygon(renderer: *const SDL_Renderer,
+                           vx: *const c_short,
+                           vy: *const c_short,
+                           n: c_int,
+                           texture: *mut SDL_Surface,
+                           texture_dx: c_int,
+                           texture_dy: c_int)
+                           -> c_int;
+    pub fn bezierColor(renderer: *const SDL_Renderer,
+                       vx: *const c_short,
+                       vy: *const c_short,
+                       n: c_int,
+                       s: c_int,
+                       color: c_uint)
+                       -> c_int;
+    pub fn bezierRGBA(renderer: *const SDL_Renderer,
+                      vx: *const c_short,
+                      vy: *const c_short,
+                      n: c_int,
+                      s: c_int,
+                      r: c_uchar,
+                      g: c_uchar,
+                      b: c_uchar,
+                      a: c_uchar)
+                      -> c_int;
+    pub fn gfxPrimitivesSetFont(fontdata: *const c_void, cw: c_uint, ch: c_uint);
+    pub fn gfxPrimitivesSetFontRotation(rotation: c_uint);
+    pub fn characterColor(renderer: *const SDL_Renderer,
+                          x: c_short,
+                          y: c_short,
+                          c: c_char,
+                          color: c_uint)
+                          -> c_int;
+    pub fn characterRGBA(renderer: *const SDL_Renderer,
+                         x: c_short,
+                         y: c_short,
+                         c: c_char,
+                         r: c_uchar,
+                         g: c_uchar,
+                         b: c_uchar,
+                         a: c_uchar)
+                         -> c_int;
+    pub fn stringColor(renderer: *const SDL_Renderer,
+                       x: c_short,
+                       y: c_short,
+                       s: *mut c_char,
+                       color: c_uint)
+                       -> c_int;
+    pub fn stringRGBA(renderer: *const SDL_Renderer,
+                      x: c_short,
+                      y: c_short,
+                      s: *mut c_char,
+                      r: c_uchar,
+                      g: c_uchar,
+                      b: c_uchar,
+                      a: c_uchar)
+                      -> c_int;
+}
