+++ src/sdl2/gfx/imagefilter.rs
@@ -4,100 +4,21 @@ use std::mem;
 use libc::{self,size_t, c_void, c_uint, c_int};
 use ::get_error;
 use c_vec::CVec;
-
-mod ll {
-    /* automatically generated by rust-bindgen */
-
-    use libc::*;
-    extern "C" {
-        pub fn SDL_imageFilterMMXdetect() -> c_int;
-        pub fn SDL_imageFilterMMXoff();
-        pub fn SDL_imageFilterMMXon();
-        pub fn SDL_imageFilterAdd(Src1: *mut u8, Src2: *mut u8,
-                                  Dest: *mut u8, length: c_uint) -> c_int;
-        pub fn SDL_imageFilterMean(Src1: *mut u8, Src2: *mut u8,
-                                   Dest: *mut u8, length: c_uint) -> c_int;
-        pub fn SDL_imageFilterSub(Src1: *mut u8, Src2: *mut u8,
-                                  Dest: *mut u8, length: c_uint) -> c_int;
-        pub fn SDL_imageFilterAbsDiff(Src1: *mut u8, Src2: *mut u8,
-                                      Dest: *mut u8, length: c_uint) ->
-            c_int;
-        pub fn SDL_imageFilterMult(Src1: *mut u8, Src2: *mut u8,
-                                   Dest: *mut u8, length: c_uint) -> c_int;
-        pub fn SDL_imageFilterMultNor(Src1: *mut u8, Src2: *mut u8,
-                                      Dest: *mut u8, length: c_uint) ->
-            c_int;
-        pub fn SDL_imageFilterMultDivby2(Src1: *mut u8, Src2: *mut u8,
-                                         Dest: *mut u8, length: c_uint) ->
-            c_int;
-        pub fn SDL_imageFilterMultDivby4(Src1: *mut u8, Src2: *mut u8,
-                                         Dest: *mut u8, length: c_uint) ->
-            c_int;
-        pub fn SDL_imageFilterBitAnd(Src1: *mut u8, Src2: *mut u8,
-                                     Dest: *mut u8, length: c_uint) -> c_int;
-        pub fn SDL_imageFilterBitOr(Src1: *mut u8, Src2: *mut u8,
-                                    Dest: *mut u8, length: c_uint) -> c_int;
-        pub fn SDL_imageFilterDiv(Src1: *mut u8, Src2: *mut u8,
-                                  Dest: *mut u8, length: c_uint) -> c_int;
-        pub fn SDL_imageFilterBitNegation(Src1: *mut u8, Dest: *mut u8,
-                                          length: c_uint) -> c_int;
-        pub fn SDL_imageFilterAddByte(Src1: *mut u8, Dest: *mut u8,
-                                      length: c_uint, C: u8) -> c_int;
-        pub fn SDL_imageFilterAddUint(Src1: *mut u8, Dest: *mut u8,
-                                      length: c_uint, C: c_uint) -> c_int;
-        pub fn SDL_imageFilterAddByteToHalf(Src1: *mut u8,
-                                            Dest: *mut u8, length: c_uint,
-                                            C: u8) -> c_int;
-        pub fn SDL_imageFilterSubByte(Src1: *mut u8, Dest: *mut u8,
-                                      length: c_uint, C: u8) -> c_int;
-        pub fn SDL_imageFilterSubUint(Src1: *mut u8, Dest: *mut u8,
-                                      length: c_uint, C: c_uint) -> c_int;
-        pub fn SDL_imageFilterShiftRight(Src1: *mut u8, Dest: *mut u8,
-                                         length: c_uint, N: u8) -> c_int;
-        pub fn SDL_imageFilterShiftRightUint(Src1: *mut u8,
-                                             Dest: *mut u8, length: c_uint,
-                                             N: u8) -> c_int;
-        pub fn SDL_imageFilterMultByByte(Src1: *mut u8, Dest: *mut u8,
-                                         length: c_uint, C: u8) -> c_int;
-        pub fn SDL_imageFilterShiftRightAndMultByByte(Src1: *mut u8,
-                                                      Dest: *mut u8,
-                                                      length: c_uint, N: u8,
-                                                      C: u8) -> c_int;
-        pub fn SDL_imageFilterShiftLeftByte(Src1: *mut u8,
-                                            Dest: *mut u8, length: c_uint,
-                                            N: u8) -> c_int;
-        pub fn SDL_imageFilterShiftLeftUint(Src1: *mut u8,
-                                            Dest: *mut u8, length: c_uint,
-                                            N: u8) -> c_int;
-        pub fn SDL_imageFilterShiftLeft(Src1: *mut u8, Dest: *mut u8,
-                                        length: c_uint, N: u8) -> c_int;
-        pub fn SDL_imageFilterBinarizeUsingThreshold(Src1: *mut u8,
-                                                     Dest: *mut u8,
-                                                     length: c_uint, T: u8)
-                                                     -> c_int;
-        pub fn SDL_imageFilterClipToRange(Src1: *mut u8, Dest: *mut u8,
-                                          length: c_uint, Tmin: u8,
-                                          Tmax: u8) -> c_int;
-        pub fn SDL_imageFilterNormalizeLinear(Src: *mut u8,
-                                              Dest: *mut u8, length: c_uint,
-                                              Cmin: c_int, Cmax: c_int,
-                                              Nmin: c_int, Nmax: c_int) -> c_int;
-    }
-}
+use sys::gfx::imagefilter;
 
 /// MMX detection routine (with override flag).
 pub fn mmx_detect() -> bool {
-    unsafe { ll::SDL_imageFilterMMXdetect() == 1 }
+    unsafe { imagefilter::SDL_imageFilterMMXdetect() == 1 }
 }
 
 /// Disable MMX check for filter functions and and force to use non-MMX C based code.
 pub fn mmx_off() {
-    unsafe { ll::SDL_imageFilterMMXoff() }
+    unsafe { imagefilter::SDL_imageFilterMMXoff() }
 }
 
 /// Enable MMX check for filter functions and use MMX code if available.
 pub fn mmx_on() {
-    unsafe { ll::SDL_imageFilterMMXon() }
+    unsafe { imagefilter::SDL_imageFilterMMXon() }
 }
 
 #[inline]
@@ -115,7 +36,7 @@ pub fn add(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterAdd(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterAdd(mem::transmute(src1.get(0)),
                                               mem::transmute(src2.get(0)),
                                               mem::transmute(dest.get(0)),
                                               size as c_uint) };
@@ -128,7 +49,7 @@ pub fn mean(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterMean(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterMean(mem::transmute(src1.get(0)),
                                                mem::transmute(src2.get(0)),
                                                mem::transmute(dest.get(0)),
                                                size as c_uint) };
@@ -141,7 +62,7 @@ pub fn sub(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterSub(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterSub(mem::transmute(src1.get(0)),
                                               mem::transmute(src2.get(0)),
                                               mem::transmute(dest.get(0)),
                                               size as c_uint) };
@@ -154,7 +75,7 @@ pub fn abs_diff(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterAbsDiff(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterAbsDiff(mem::transmute(src1.get(0)),
                                                   mem::transmute(src2.get(0)),
                                                   mem::transmute(dest.get(0)),
                                                   size as c_uint) };
@@ -167,7 +88,7 @@ pub fn mult(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterMult(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterMult(mem::transmute(src1.get(0)),
                                                mem::transmute(src2.get(0)),
                                                mem::transmute(dest.get(0)),
                                                size as c_uint) };
@@ -180,7 +101,7 @@ pub fn mult_nor(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterMultNor(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterMultNor(mem::transmute(src1.get(0)),
                                                   mem::transmute(src2.get(0)),
                                                   mem::transmute(dest.get(0)),
                                                   size as c_uint) };
@@ -193,7 +114,7 @@ pub fn mult_div_by2(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> 
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterMultDivby2(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterMultDivby2(mem::transmute(src1.get(0)),
                                                      mem::transmute(src2.get(0)),
                                                      mem::transmute(dest.get(0)),
                                                      size as c_uint) };
@@ -206,7 +127,7 @@ pub fn mult_div_by4(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> 
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterMultDivby4(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterMultDivby4(mem::transmute(src1.get(0)),
                                                      mem::transmute(src2.get(0)),
                                                      mem::transmute(dest.get(0)),
                                                      size as c_uint) };
@@ -219,7 +140,7 @@ pub fn bit_and(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterBitAnd(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterBitAnd(mem::transmute(src1.get(0)),
                                                  mem::transmute(src2.get(0)),
                                                  mem::transmute(dest.get(0)),
                                                  size as c_uint) };
@@ -232,7 +153,7 @@ pub fn bit_or(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterBitOr(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterBitOr(mem::transmute(src1.get(0)),
                                                 mem::transmute(src2.get(0)),
                                                 mem::transmute(dest.get(0)),
                                                 size as c_uint) };
@@ -245,7 +166,7 @@ pub fn div(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
     assert_eq!(src1.len(), src2.len());
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterDiv(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterDiv(mem::transmute(src1.get(0)),
                                               mem::transmute(src2.get(0)),
                                               mem::transmute(dest.get(0)),
                                               size as c_uint) };
@@ -257,7 +178,7 @@ pub fn div(src1: CVec<u8>, src2: CVec<u8>) -> Result<CVec<u8>, String> {
 pub fn bit_negation(src1: CVec<u8>) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterBitNegation(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterBitNegation(mem::transmute(src1.get(0)),
                                                       mem::transmute(dest.get(0)),
                                                       size as c_uint) };
     if ret == 0 { Ok(dest) }
@@ -268,7 +189,7 @@ pub fn bit_negation(src1: CVec<u8>) -> Result<CVec<u8>, String> {
 pub fn add_byte(src1: CVec<u8>, c: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterAddByte(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterAddByte(mem::transmute(src1.get(0)),
                                                   mem::transmute(dest.get(0)),
                                                   size as c_uint, c) };
     if ret == 0 { Ok(dest) }
@@ -279,7 +200,7 @@ pub fn add_byte(src1: CVec<u8>, c: u8) -> Result<CVec<u8>, String> {
 pub fn add_uint(src1: CVec<u8>, c: u32) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterAddUint(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterAddUint(mem::transmute(src1.get(0)),
                                                   mem::transmute(dest.get(0)),
                                                   size as c_uint, c) };
     if ret == 0 { Ok(dest) }
@@ -290,7 +211,7 @@ pub fn add_uint(src1: CVec<u8>, c: u32) -> Result<CVec<u8>, String> {
 pub fn add_byte_to_half(src1: CVec<u8>, c: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterAddByteToHalf(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterAddByteToHalf(mem::transmute(src1.get(0)),
                                                         mem::transmute(dest.get(0)),
                                                         size as c_uint, c) };
     if ret == 0 { Ok(dest) }
@@ -301,7 +222,7 @@ pub fn add_byte_to_half(src1: CVec<u8>, c: u8) -> Result<CVec<u8>, String> {
 pub fn sub_byte(src1: CVec<u8>, c: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterSubByte(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterSubByte(mem::transmute(src1.get(0)),
                                                   mem::transmute(dest.get(0)),
                                                   size as c_uint, c) };
     if ret == 0 { Ok(dest) }
@@ -312,7 +233,7 @@ pub fn sub_byte(src1: CVec<u8>, c: u8) -> Result<CVec<u8>, String> {
 pub fn sub_uint(src1: CVec<u8>, c: u32) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterSubUint(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterSubUint(mem::transmute(src1.get(0)),
                                                   mem::transmute(dest.get(0)),
                                                   size as c_uint, c) };
     if ret == 0 { Ok(dest) }
@@ -323,7 +244,7 @@ pub fn sub_uint(src1: CVec<u8>, c: u32) -> Result<CVec<u8>, String> {
 pub fn shift_right(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterShiftRight(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterShiftRight(mem::transmute(src1.get(0)),
                                                      mem::transmute(dest.get(0)),
                                                      size as c_uint, n) };
     if ret == 0 { Ok(dest) }
@@ -334,7 +255,7 @@ pub fn shift_right(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
 pub fn shift_right_uint(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterShiftRightUint(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterShiftRightUint(mem::transmute(src1.get(0)),
                                                          mem::transmute(dest.get(0)),
                                                          size as c_uint, n) };
     if ret == 0 { Ok(dest) }
@@ -345,7 +266,7 @@ pub fn shift_right_uint(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
 pub fn mult_by_byte(src1: CVec<u8>, c: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterMultByByte(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterMultByByte(mem::transmute(src1.get(0)),
                                                      mem::transmute(dest.get(0)),
                                                      size as c_uint, c) };
     if ret == 0 { Ok(dest) }
@@ -356,7 +277,7 @@ pub fn mult_by_byte(src1: CVec<u8>, c: u8) -> Result<CVec<u8>, String> {
 pub fn shift_right_and_mult_by_byte(src1: CVec<u8>, n: u8, c: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterShiftRightAndMultByByte(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterShiftRightAndMultByByte(mem::transmute(src1.get(0)),
                                                                   mem::transmute(dest.get(0)),
                                                                   size as c_uint, n, c) };
     if ret == 0 { Ok(dest) }
@@ -367,7 +288,7 @@ pub fn shift_right_and_mult_by_byte(src1: CVec<u8>, n: u8, c: u8) -> Result<CVec
 pub fn shift_left_byte(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterShiftLeftByte(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterShiftLeftByte(mem::transmute(src1.get(0)),
                                                         mem::transmute(dest.get(0)),
                                                         size as c_uint, n) };
     if ret == 0 { Ok(dest) }
@@ -378,7 +299,7 @@ pub fn shift_left_byte(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
 pub fn shift_left_uint(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterShiftLeftUint(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterShiftLeftUint(mem::transmute(src1.get(0)),
                                                         mem::transmute(dest.get(0)),
                                                         size as c_uint, n) };
     if ret == 0 { Ok(dest) }
@@ -389,7 +310,7 @@ pub fn shift_left_uint(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
 pub fn shift_left(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterShiftLeft(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterShiftLeft(mem::transmute(src1.get(0)),
                                                     mem::transmute(dest.get(0)),
                                                     size as c_uint, n) };
     if ret == 0 { Ok(dest) }
@@ -400,7 +321,7 @@ pub fn shift_left(src1: CVec<u8>, n: u8) -> Result<CVec<u8>, String> {
 pub fn binarize_using_threshold(src1: CVec<u8>, t: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterBinarizeUsingThreshold(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterBinarizeUsingThreshold(mem::transmute(src1.get(0)),
                                                                  mem::transmute(dest.get(0)),
                                                                  size as c_uint, t) };
     if ret == 0 { Ok(dest) }
@@ -411,7 +332,7 @@ pub fn binarize_using_threshold(src1: CVec<u8>, t: u8) -> Result<CVec<u8>, Strin
 pub fn clip_to_range(src1: CVec<u8>, tmin: u8, tmax: u8) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterClipToRange(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterClipToRange(mem::transmute(src1.get(0)),
                                                       mem::transmute(dest.get(0)),
                                                       size as c_uint, tmin, tmax) };
     if ret == 0 { Ok(dest) }
@@ -422,7 +343,7 @@ pub fn clip_to_range(src1: CVec<u8>, tmin: u8, tmax: u8) -> Result<CVec<u8>, Str
 pub fn normalize_linear(src1: CVec<u8>, cmin: i32, cmax: i32, nmin: i32, nmax: i32) -> Result<CVec<u8>, String> {
     let size = src1.len();
     let dest = cvec_with_size(size);
-    let ret = unsafe { ll::SDL_imageFilterNormalizeLinear(mem::transmute(src1.get(0)),
+    let ret = unsafe { imagefilter::SDL_imageFilterNormalizeLinear(mem::transmute(src1.get(0)),
                                                           mem::transmute(dest.get(0)),
                                                           size as c_uint,
                                                           cmin as c_int, cmax as c_int,
