+++ src/sdl2/gfx/mod.rs
@@ -20,10 +20,6 @@
 //! features = ["gfx"]
 //! ```
 
-// Setup linking for all targets.
-#[link(name="SDL2_gfx")]
-extern {}
-
 pub mod primitives;
 pub mod rotozoom;
 pub mod framerate;
