+++ src/sdl2/gfx/rotozoom.rs
@@ -4,37 +4,8 @@ use libc::c_int;
 use ::surface::Surface;
 use ::get_error;
 pub use std::f64::consts::PI;
+use sys::gfx::rotozoom;
 
-
-mod ll {
-    /* automatically generated by rust-bindgen */
-
-    use libc::*;
-    use sys::SDL_Surface;
-    extern "C" {
-        pub fn rotozoomSurface(src: *mut SDL_Surface, angle: c_double,
-                               zoom: c_double, smooth: c_int) -> *mut SDL_Surface;
-        pub fn rotozoomSurfaceXY(src: *mut SDL_Surface, angle: c_double,
-                                 zoomx: c_double, zoomy: c_double, smooth: c_int)
-                                 -> *mut SDL_Surface;
-        pub fn rotozoomSurfaceSize(width: c_int, height: c_int, angle: c_double,
-                                   zoom: c_double, dstwidth: *mut c_int,
-                                   dstheight: *mut c_int);
-        pub fn rotozoomSurfaceSizeXY(width: c_int, height: c_int, angle: c_double,
-                                     zoomx: c_double, zoomy: c_double,
-                                     dstwidth: *mut c_int, dstheight: *mut c_int);
-        pub fn zoomSurface(src: *mut SDL_Surface, zoomx: c_double,
-                           zoomy: c_double, smooth: c_int) -> *mut SDL_Surface;
-        pub fn zoomSurfaceSize(width: c_int, height: c_int, zoomx: c_double,
-                               zoomy: c_double, dstwidth: *mut c_int,
-                               dstheight: *mut c_int);
-        pub fn shrinkSurface(src: *mut SDL_Surface, factorx: c_int,
-                             factory: c_int) -> *mut SDL_Surface;
-        pub fn rotateSurface90Degrees(src: *mut SDL_Surface,
-                                      numClockwiseTurns: c_int) ->
-            *mut SDL_Surface;
-    }
-}
 
 /// `RotozoomSurface` for work with rust-sdl2 Surface type
 pub trait RotozoomSurface {
@@ -53,7 +24,7 @@ pub trait RotozoomSurface {
 impl<'a> RotozoomSurface for Surface<'a> {
     fn rotozoom(&self, angle: f64, zoom: f64, smooth: bool) -> Result<Surface, String> {
         let raw = unsafe {
-            ll::rotozoomSurface(self.raw(), angle, zoom, smooth as c_int)
+            rotozoom::rotozoomSurface(self.raw(), angle, zoom, smooth as c_int)
         };
         if (raw as *mut ()).is_null() {
             Err(get_error())
@@ -63,7 +34,7 @@ impl<'a> RotozoomSurface for Surface<'a> {
     }
     fn rotozoom_xy(&self, angle: f64, zoomx: f64, zoomy: f64, smooth: bool) -> Result<Surface, String> {
         let raw = unsafe {
-            ll::rotozoomSurfaceXY(self.raw(), angle, zoomx, zoomy, smooth as c_int)
+            rotozoom::rotozoomSurfaceXY(self.raw(), angle, zoomx, zoomy, smooth as c_int)
         };
         if (raw as *mut ()).is_null() {
             Err(get_error())
@@ -73,7 +44,7 @@ impl<'a> RotozoomSurface for Surface<'a> {
     }
     fn zoom(&self, zoomx: f64, zoomy: f64, smooth: bool) -> Result<Surface, String> {
         let raw = unsafe {
-            ll::zoomSurface(self.raw(), zoomx, zoomy, smooth as c_int)
+            rotozoom::zoomSurface(self.raw(), zoomx, zoomy, smooth as c_int)
         };
         if (raw as *mut ()).is_null() {
             Err(get_error())
@@ -83,7 +54,7 @@ impl<'a> RotozoomSurface for Surface<'a> {
     }
     fn shrink(&self, factorx: i32, factory: i32) -> Result<Surface, String> {
         let raw = unsafe {
-            ll::shrinkSurface(self.raw(), factorx as c_int, factory as c_int)
+            rotozoom::shrinkSurface(self.raw(), factorx as c_int, factory as c_int)
         };
         if (raw as *mut ()).is_null() {
             Err(get_error())
@@ -93,7 +64,7 @@ impl<'a> RotozoomSurface for Surface<'a> {
     }
     fn rotate_90deg(&self, turns: i32) -> Result<Surface, String> {
         let raw = unsafe {
-            ll::rotateSurface90Degrees(self.raw(), turns as c_int)
+            rotozoom::rotateSurface90Degrees(self.raw(), turns as c_int)
         };
         if (raw as *mut ()).is_null() {
             Err(get_error())
@@ -106,20 +77,20 @@ impl<'a> RotozoomSurface for Surface<'a> {
 pub fn get_zoom_size(width: i32, height: i32, zoomx: f64, zoomy: f64) -> (i32, i32) {
     let mut w: c_int = 0;
     let mut h: c_int = 0;
-    unsafe { ll::zoomSurfaceSize(width as c_int, height as c_int, zoomx, zoomy, &mut w, &mut h) }
+    unsafe { rotozoom::zoomSurfaceSize(width as c_int, height as c_int, zoomx, zoomy, &mut w, &mut h) }
     (w as i32, h as i32)
 }
 
 pub fn get_rotozoom_size(width: i32, height: i32, angle: f64, zoom: f64) -> (i32, i32) {
     let mut w: c_int = 0;
     let mut h: c_int = 0;
-    unsafe { ll::rotozoomSurfaceSize(width as c_int, height as c_int, angle, zoom, &mut w, &mut h) }
+    unsafe { rotozoom::rotozoomSurfaceSize(width as c_int, height as c_int, angle, zoom, &mut w, &mut h) }
     (w as i32, h as i32)
 }
 
 pub fn get_rotozoom_xy_size(width: i32, height: i32, angle: f64, zoomx: f64, zoomy: f64) -> (i32, i32) {
     let mut w: c_int = 0;
     let mut h: c_int = 0;
-    unsafe { ll::rotozoomSurfaceSizeXY(width as c_int, height as c_int, angle, zoomx, zoomy, &mut w, &mut h) }
+    unsafe { rotozoom::rotozoomSurfaceSizeXY(width as c_int, height as c_int, angle, zoomx, zoomy, &mut w, &mut h) }
     (w as i32, h as i32)
 }
